Changed datatype to Float from Decimal

+++ CSharp/MotionController/MotionController.Sensor.Db.Data/Models/GameSessionBallPosition.cs
@@ -17,13 +17,13 @@ public sealed class GameSessionBallPosition : DatabaseModel
     public int GameSessionId { get; set; }
 
     [Column("X")]
-    public decimal X { get; set; }
+    public float X { get; set; }
 
     [Column("Y")]
-    public decimal Y { get; set; }
+    public float Y { get; set; }
 
     [Column("Z")]
-    public decimal Z { get; set; }
+    public float Z { get; set; }
 
     [Column("Created")]
     public DateTime Created { get; set; }
